Split Rss09xParser and Rss20Parser to separate files
