+++ rss/rss09xparser.cpp
@@ -1,4 +1,4 @@
-#include "rssppinternal.h"
+#include "rss09xparser.h"
 
 #include <cstring>
 
@@ -11,21 +11,6 @@
 using namespace newsboat;
 
 namespace rsspp {
-
-void Rss20Parser::parse_feed(Feed& f, xmlNode* rootNode)
-{
-	if (!rootNode)
-		throw Exception(_("XML root node is NULL"));
-
-	if (rootNode->ns) {
-		const char* ns = (const char*)rootNode->ns->href;
-		if (strcmp(ns, RSS20USERLAND_URI) == 0) {
-			this->ns = strdup(ns);
-		}
-	}
-
-	Rss09xParser::parse_feed(f, rootNode);
-}
 
 void Rss09xParser::parse_feed(Feed& f, xmlNode* rootNode)
 {
